requirements.txt
matplotlib==3.10.1
mysql-connector-python==9.3.0
